+++ src/elements/mod.rs
@@ -1,6 +1,7 @@
 mod b_block;
 mod b_box;
 mod b_button;
+mod b_icon;
 mod b_notification;
 mod b_table;
 mod b_title;
@@ -8,6 +9,7 @@ mod b_title;
 pub use b_block::*;
 pub use b_box::*;
 pub use b_button::*;
+pub use b_icon::*;
 pub use b_notification::*;
 pub use b_table::*;
 pub use b_title::*;